handle incorrect input requests

+++ models/base.py
@@ -6,7 +6,7 @@ def commit(obj):
     Function for convenient commit
     """
     db.session.add(obj)
-    db.session.commit(obj)
+    db.session.commit()
     db.session.refresh(obj)
     return obj
 
@@ -39,6 +39,13 @@ class Model(object):
 
     @classmethod
     def delete(cls, row_id):
+        """
+        Delete record by id
+
+        cls: class
+        row_id: record id
+        return: int (1 if deleted else 0)
+        """
         obj = cls.query.filter_by(id=row_id).first()
         try:
             db.session.delete(obj)
@@ -56,10 +63,40 @@ class Model(object):
         row_id: record id
         rel_obj: related object
         """
-        obj = db.query.filter_by(id=row_id).first()
+        obj = cls.query.filter_by(id=row_id).first()
         if cls.__name__ == 'Actor':
             obj.filmography.append(rel_obj)
         elif cls.__name__ == 'Movie':
             obj.cast.append(rel_obj)
         return commit(obj)
 
+    @classmethod
+    def remove_relation(cls, row_id, rel_obj):
+        """
+        Remove certain relation
+
+        cls: class
+        row_id: record id
+        rel_obj: related object
+        """
+        obj = cls.query.filter_by(id=row_id).first()
+        if cls.__name__ == 'Actor':
+            obj.filmography.delete(rel_obj)
+        elif cls.__name__ == 'Movie':
+            obj.cast.delete(rel_obj)
+        return commit(obj)
+
+    @classmethod
+    def clear_relations(cls, row_id):
+        """
+        Remove all relations by id
+
+        cls: class
+        row_id: record id
+        """
+        obj = cls.query.filter_by(id=row_id).first()
+        if cls.__name__ == 'Actor':
+            obj.filmography.clear()
+        elif cls.__name__ == 'Movie':
+            obj.cast.clear()
+        return commit(obj)
